archive/requirements.txt
apscheduler
beautifulsoup4
cohere
diskcache
feedparser
html5lib
matplotlib>=3.7.0
networkx
numpy
openai>=1.0.0
pandas
plotly>=5.17.0
pulp
python-dotenv
pytest
requests
scikit-learn
selenium
shap
sqlalchemy
streamlit>=1.18
streamlit-extras
thefuzz
undetected-chromedriver
scrapy
streamlit-aggrid